MeasureMe v2: xG integration, Understat fix, continuous scoring, config update
- Fix Understat scraper: switch from broken HTML scraping to AJAX API
  (/getLeagueData endpoint), rewrite name map to match DB short names,
  batch fixture lookup (6s vs 15min). 1240/1241 matches matched.
- Add xG shock multiplier to MeasureMe grid search (same formula as
  pricing-engine: clamp(1+0.3*(teamXg-oppXg)*sign(gd), 0.4, 1.8))
- New winner: K=20, decay=0.001, zp=800 (was K=40 without xG)
- Continuous scoring for all 7 indices (no more step functions)
- Update pricing engine config: SHOCK_K=20, CARRY_DECAY=0.001,
  PRICE_ZERO=800 (was 1000 hardcoded), add PRICE_ZERO constant
- MeasureMe frontend: remove slope column, add slope dropdown,
  handle decimal scores from NUMERIC columns
- DB schema: score columns INTEGER → NUMERIC
- Add measureme-validate.ts diagnostic script

+++ scheduler/sql/recreate_measureme_results.sql
@@ -11,7 +11,7 @@ CREATE TABLE measureme_results (
   k_factor NUMERIC NOT NULL,
   decay NUMERIC NOT NULL,
   zero_point NUMERIC NOT NULL,
-  composite_score INTEGER NOT NULL,
+  composite_score NUMERIC NOT NULL,
 
   -- Raw index values
   surprise_r2 NUMERIC,
@@ -22,14 +22,14 @@ CREATE TABLE measureme_results (
   mean_rev_sharpe NUMERIC,
   info_ratio NUMERIC,
 
-  -- Index scores (0-100)
-  surprise_r2_score INTEGER,
-  drift_score INTEGER,
-  floor_hit_score INTEGER,
-  kurtosis_score INTEGER,
-  vol_uni_score INTEGER,
-  mean_rev_score INTEGER,
-  info_score INTEGER,
+  -- Index scores (0-100, continuous)
+  surprise_r2_score NUMERIC,
+  drift_score NUMERIC,
+  floor_hit_score NUMERIC,
+  kurtosis_score NUMERIC,
+  vol_uni_score NUMERIC,
+  mean_rev_score NUMERIC,
+  info_score NUMERIC,
 
   -- Summary stats
   avg_match_move_pct NUMERIC,
